Reorganize a bit web/instances
